backend/requirements.txt
django==3.2.7
channels==3.0.2
django-apscheduler==0.6.2
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-allauth==0.45.0
django-nose==1.4.7
django-smtp-ssl==1.0
drf-yasg2==1.19.4
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
bs4==0.0.1
pytube==12.0.0
lxml==4.9.1
user-agents==2.2.0